Remove all the compile macros from API header files
