+++ .ycm_extra_conf.py
@@ -51,8 +51,8 @@ flags = [
 '-Winline',
 '-Wdisabled-optimization',
 '-DNDEBUG',
-'-D_POSIX_C_SOURCE=200809L',
-'-D_BSD_SOURCE',
+#'-D_POSIX_C_SOURCE=200809L',
+#'-D_BSD_SOURCE',
 # THIS IS IMPORTANT! Without a "-std=<something>" flag, clang won't know which
 # language to use when compiling headers. So it will guess. Badly. So C++
 # headers will be compiled as C headers. You don't want that so ALWAYS specify
