+++ fev/fev_timer_service.c
@@ -34,11 +34,14 @@
  * =============================================================================
  */
 
+#ifndef  _POSIX_C_SOURCE
+# define _POSIX_C_SOURCE 200809L
+#endif
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <errno.h>
-#include <time.h>
 #include <limits.h>
 
 #include "fev_tmsvc_modules.h"
