+++ include/flibs/fco.h
@@ -1,6 +1,15 @@
 #ifndef FCOROUTINE_H
 #define FCOROUTINE_H
 
+/**
+ * fco context, to help user make huge amount of user contexts, also provide
+ *  the basic interfaces like yield/resume to build the application easier.
+ *
+ * Notes: This fco_context is not recommended to use, since it may leading
+ *  portability issues. And in certain 64bit platform may be not working
+ *  as expected.
+ */
+
 #ifdef __cplusplus
 extern "C" {
 #endif
